astro_profiling/requirements.txt
prokerala_api
fastapi
uvicorn
python-dotenv
pydantic